poker in new file, not done and SP buggy
